Fixes, bump to version 1.2.0
Fixes #13

We now hook openCV's error handling function to prevent any action globally.
We now ensure that any image decoded is valid. If it's not valid, we return false / 0.d for public classifier methods.
Removed exception throwing from the C interface because I was dumb to put it there in the first place.
Explicitly disable OpenCL in OpenCV
Upgrade to OpenCV 3.4.1
Added the ability to load prototxt and model from memory rather than having to supply a file path to the library.